# producer-service/src/main/java/ru/t1/owner/dto/UpdateOwnerDto.java
package ru.t1.owner.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

@Getter
@Setter
@ToString
@NoArgsConstructor
public class UpdateOwnerDto {
    private String name;
    private String email;
}