# producer-service/src/main/java/ru/t1/solution/dto/UpdateSolutionDto.java
package ru.t1.solution.dto;

import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Positive;
import jakarta.validation.constraints.Size;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class UpdateSolutionDto {
    @Size(min = 2, max = 30, message = "validation name size error")
    @Pattern(regexp = "^(?=.*[a-zA-Zа-яёЁА-Я\\d_\\S]).+$")
    private String name;
    @Builder.Default
    private String version = "0.0.1";
    @Positive
    private Long ownerId;
}